Visualization and Actual Use
Added a nice way of visualizing a user relative to their peers/followers while also adding ways to get other info. Is rudimentary but I really liked looking at the graphs.
This program expects to be able to read a GitHub Access Token in a file at the start, you can also just take it out and put your own token in.
This programs visualization doesn't work with users with too many followers as the data limit will easily be reached.